move auth entities to authentication package
