+++ apps/modernization-api/src/test/java/gov/cdc/nbs/AuthenticationSteps.java
@@ -16,9 +16,8 @@ import org.springframework.test.web.servlet.MockMvc;
 import org.springframework.test.web.servlet.request.MockMvcRequestBuilders;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
-
+import gov.cdc.nbs.authentication.AuthUserRepository;
 import gov.cdc.nbs.model.LoginRequest;
-import gov.cdc.nbs.repository.AuthUserRepository;
 import gov.cdc.nbs.support.UserMother;
 import io.cucumber.java.Before;
 import io.cucumber.java.en.Given;
@@ -58,10 +57,10 @@ public class AuthenticationSteps {
     public void i_have_authenticated_as_a_user() throws Exception {
         var user = TestContext.user;
         var response = mvc.perform(MockMvcRequestBuilders
-                        .post("/login")
-                        .content(mapper.writeValueAsBytes(new LoginRequest(user.getUserId(), "")))
-                        .contentType(MediaType.APPLICATION_JSON)
-                        .accept(MediaType.APPLICATION_JSON))
+                .post("/login")
+                .content(mapper.writeValueAsBytes(new LoginRequest(user.getUserId(), "")))
+                .contentType(MediaType.APPLICATION_JSON)
+                .accept(MediaType.APPLICATION_JSON))
                 .andExpect(status().isOk()).andReturn();
         var tokenCookie = response.getResponse().getCookie("nbs_token");
         assertNotNull(tokenCookie);
