+++ apps/modernization-api/src/test/java/gov/cdc/nbs/FindUsersSteps.java
@@ -14,15 +14,14 @@ import org.springframework.data.domain.Page;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.test.context.ActiveProfiles;
 import org.springframework.test.context.junit4.SpringRunner;
-
+import gov.cdc.nbs.authentication.AuthPermSet;
+import gov.cdc.nbs.authentication.AuthPermSetRepository;
+import gov.cdc.nbs.authentication.AuthUser;
+import gov.cdc.nbs.authentication.AuthUserRepository;
+import gov.cdc.nbs.authentication.AuthUserRole;
+import gov.cdc.nbs.authentication.AuthUserRoleRepository;
 import gov.cdc.nbs.controller.UserController;
 import gov.cdc.nbs.entity.enums.RecordStatus;
-import gov.cdc.nbs.entity.odse.AuthPermSet;
-import gov.cdc.nbs.entity.odse.AuthUser;
-import gov.cdc.nbs.entity.odse.AuthUserRole;
-import gov.cdc.nbs.repository.AuthPermSetRepository;
-import gov.cdc.nbs.repository.AuthUserRepository;
-import gov.cdc.nbs.repository.AuthUserRoleRepository;
 import gov.cdc.nbs.support.PermissionMother;
 import gov.cdc.nbs.support.UserMother;
 import io.cucumber.java.Before;
