+++ apps/question-bank/src/main/resources/application.yml
@@ -6,7 +6,7 @@ nbs:
     tokenIssuer: gov.cdc.nbs
     tokenExpirationMillis: 1800000 # 30 minutes
     max-page-size: 50
- 
+
 spring:
   graphql:
     path: /graphql
@@ -28,7 +28,7 @@ spring:
         format_sql: true
         default_schema: dbo
     hibernate:
-      ddl-auto: validate
+      ddl-auto: none
   datasource:
     url: jdbc:sqlserver://${nbs.datasource.server}:1433;database=question_bank;encrypt=true;trustServerCertificate=true;
     username: qb_admin
