+++ apps/question-bank/src/test/resources/application-test.yml
@@ -1,6 +1,11 @@
 nbs:
   datasource:
     server: localhost
+  security:
+    tokenSecret: qJ5+oDurrJnxfaAz6h4Eg1saP7NSeHuQQ/Yq4vh+mbi4B4QYThc/M+33VtPJ5O2A
+    tokenIssuer: gov.cdc.nbs
+    tokenExpirationMillis: 1800000 # 30 minutes
+    max-page-size: 50
 
 spring:
   flyway:
@@ -18,7 +23,7 @@ spring:
         format_sql: true
         default_schema: DBO
     hibernate:
-      ddl-auto: validate
+      ddl-auto: none
   datasource:
     driver-class-name: org.h2.Driver
     url: jdbc:h2:mem:question_bank;MODE=MSSQLServer;DB_CLOSE_DELAY=-1;
