+++ libs/database-entities/src/main/java/gov/cdc/nbs/authentication/AuthBusObjRt.java
@@ -1,4 +1,4 @@
-package gov.cdc.nbs.entity.odse;
+package gov.cdc.nbs.authentication;
 
 import java.time.Instant;
 
@@ -6,10 +6,15 @@ import javax.persistence.Column;
 import javax.persistence.Entity;
 import javax.persistence.EnumType;
 import javax.persistence.Enumerated;
+import javax.persistence.FetchType;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
+import javax.persistence.JoinColumn;
+import javax.persistence.ManyToOne;
 import javax.persistence.Table;
+
+import com.fasterxml.jackson.annotation.JsonIgnore;
 
 import gov.cdc.nbs.entity.enums.RecordStatus;
 import lombok.AllArgsConstructor;
@@ -23,19 +28,22 @@ import lombok.Setter;
 @Getter
 @Setter
 @Entity
-@Table(name = "Auth_bus_op_type")
 @Builder
-public class AuthBusOpType {
+@Table(name = "Auth_bus_obj_rt")
+public class AuthBusObjRt {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "auth_bus_op_type_uid", nullable = false)
+    @Column(name = "auth_bus_obj_rt_uid", nullable = false)
     private Long id;
 
-    @Column(name = "bus_op_nm", length = 100)
-    private String busOpNm;
+    @JsonIgnore
+    @ManyToOne(fetch = FetchType.LAZY, optional = false)
+    @JoinColumn(name = "auth_perm_set_uid", nullable = false, insertable = false, updatable = false)
+    private AuthPermSet authPermSetUid;
 
-    @Column(name = "bus_op_disp_nm", length = 1000)
-    private String busOpDispNm;
+    @ManyToOne(fetch = FetchType.LAZY, optional = false)
+    @JoinColumn(name = "auth_bus_obj_type_uid", nullable = false)
+    private AuthBusObjType authBusObjTypeUid;
 
     @Column(name = "add_time", nullable = false)
     private Instant addTime;
@@ -55,8 +63,5 @@ public class AuthBusOpType {
 
     @Column(name = "record_status_time", nullable = false)
     private Instant recordStatusTime;
-
-    @Column(name = "operation_sequence", nullable = false)
-    private Integer operationSequence;
 
 }
