+++ libs/database-entities/src/main/java/gov/cdc/nbs/authentication/AuthBusOpType.java
@@ -1,4 +1,4 @@
-package gov.cdc.nbs.entity.odse;
+package gov.cdc.nbs.authentication;
 
 import java.time.Instant;
 
@@ -6,15 +6,10 @@ import javax.persistence.Column;
 import javax.persistence.Entity;
 import javax.persistence.EnumType;
 import javax.persistence.Enumerated;
-import javax.persistence.FetchType;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
-import javax.persistence.JoinColumn;
-import javax.persistence.ManyToOne;
 import javax.persistence.Table;
-
-import com.fasterxml.jackson.annotation.JsonIgnore;
 
 import gov.cdc.nbs.entity.enums.RecordStatus;
 import lombok.AllArgsConstructor;
@@ -28,28 +23,19 @@ import lombok.Setter;
 @Getter
 @Setter
 @Entity
+@Table(name = "Auth_bus_op_type")
 @Builder
-@Table(name = "Auth_bus_op_rt")
-public class AuthBusOpRt {
+public class AuthBusOpType {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "auth_bus_op_rt_uid", nullable = false)
+    @Column(name = "auth_bus_op_type_uid", nullable = false)
     private Long id;
 
-    @ManyToOne(fetch = FetchType.LAZY, optional = false)
-    @JoinColumn(name = "auth_bus_op_type_uid", nullable = false)
-    private AuthBusOpType authBusOpTypeUid;
+    @Column(name = "bus_op_nm", length = 100)
+    private String busOpNm;
 
-    @JsonIgnore
-    @ManyToOne(fetch = FetchType.LAZY, optional = false)
-    @JoinColumn(name = "auth_bus_obj_rt_uid", nullable = false)
-    private AuthBusObjRt authBusObjRtUid;
-
-    @Column(name = "bus_op_user_rt")
-    private Character busOpUserRt;
-
-    @Column(name = "bus_op_guest_rt")
-    private Character busOpGuestRt;
+    @Column(name = "bus_op_disp_nm", length = 1000)
+    private String busOpDispNm;
 
     @Column(name = "add_time", nullable = false)
     private Instant addTime;
@@ -69,5 +55,8 @@ public class AuthBusOpRt {
 
     @Column(name = "record_status_time", nullable = false)
     private Instant recordStatusTime;
+
+    @Column(name = "operation_sequence", nullable = false)
+    private Integer operationSequence;
 
 }
