+++ libs/database-entities/src/main/java/gov/cdc/nbs/authentication/AuthUserRepository.java
@@ -1,4 +1,4 @@
-package gov.cdc.nbs.repository;
+package gov.cdc.nbs.authentication;
 
 import java.util.List;
 import java.util.Optional;
@@ -9,8 +9,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.querydsl.QuerydslPredicateExecutor;
 import org.springframework.data.repository.query.Param;
-
-import gov.cdc.nbs.entity.odse.AuthUser;
 
 public interface AuthUserRepository extends JpaRepository<AuthUser, Long>, QuerydslPredicateExecutor<AuthUser> {
     public Optional<AuthUser> findByUserId(String userId);
