+++ libs/database-entities/src/main/java/gov/cdc/nbs/authentication/AuthUserRoleRepository.java
@@ -1,12 +1,10 @@
-package gov.cdc.nbs.repository;
+package gov.cdc.nbs.authentication;
 
 import java.util.Optional;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
-
-import gov.cdc.nbs.entity.odse.AuthUserRole;
 
 public interface AuthUserRoleRepository extends JpaRepository<AuthUserRole, Long> {
